requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
certifi==2018.10.15
chardet==3.0.4
gast==0.2.0
gdown==3.6.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
idna==2.7
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.15.1
pandas==0.23.1
praw==6.0.0
prawcore==1.0.0
protobuf==3.6.0
pydot==1.2.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.4
PyYAML==3.12
requests==2.20.1
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
tensorflow-hub==0.1.1
tensorflowjs==0.6.7
termcolor==1.1.0
tqdm==4.28.1
update-checker==0.16
urllib3==1.24.1
Werkzeug==0.14.1
